# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Jaseo Job Tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /auth/password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
        "400":
          description: Incorrect current password
...
  /pm/auth/register:
    post:
      operationId: pmRegister
      tags: [pm-auth]
      summary: Register a new marketplace user (buyer or seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PmRegisterInput"
      responses:
        "201":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmUser"
        "400":
          description: Validation error or email already taken
  /pm/auth/login:
    post:
      operationId: pmLogin
      tags: [pm-auth]
      summary: Login to marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PmLoginInput"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmUser"
        "401":
          description: Invalid credentials
...
  /pm/auth/me:
    get:
      operationId: getPmMe
      tags: [pm-auth]
      summary: Get current marketplace user
      responses:
        "200":
          description: Current marketplace user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmUser"
        "401":
          description: Not authenticated
...
  /pm/admin/stats:
    get:
      operationId: getPmAdminStats
      tags: [pm-admin]
      summary: Get platform revenue statistics (admin only)
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmAdminStats"
        "403":
          description: Forbidden
  /pm/admin/users:
    get:
      operationId: listPmUsers
      tags: [pm-admin]
      summary: List all marketplace users (admin only)
      responses:
        "200":
          description: All marketplace users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PmUser"
        "403":
          description: Forbidden
  /pm/seller/stats:
    get:
      operationId: getPmSellerStats
      tags: [pm-proposals]
      summary: Get seller dashboard stats (total sales, earnings, pending payouts)
      responses:
        "200":
          description: Seller stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmSellerStats"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        writerId:
          type: ["integer", "null"]
        writerName:
          type: ["string", "null"]
        editorId:
          type: ["integer", "null"]
        editorName:
          type: ["string", "null"]
      required:
        - id
        - username
        - role
...
    PmUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        bankName:
          type: ["string", "null"]
        bankAccount:
          type: ["string", "null"]
        accountOwner:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
    PmRegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
      required:
        - name
        - email
        - password
        - role
    PmLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
...
    PmAdminStats:
      type: object
      properties:
        totalRevenue:
          type: integer
        platformRevenue:
          type: integer
        sellerPayouts:
          type: integer
        totalTransactions:
          type: integer
        totalProposals:
          type: integer
        totalSellers:
          type: integer
        totalBuyers:
          type: integer
        pendingPayouts:
          type: integer
      required:
        - totalRevenue
        - platformRevenue
        - sellerPayouts
        - totalTransactions
        - totalProposals
        - totalSellers
        - totalBuyers
        - pendingPayouts
    PmSellerStats:
      type: object
      properties:
        totalSales:
          type: integer
        totalEarnings:
          type: integer
        availableBalance:
          type: integer
        pendingPayouts:
          type: integer
        totalProposals:
          type: integer
      required:
        - totalSales
        - totalEarnings
        - availableBalance
        - pendingPayouts
        - totalProposals
...